requirements.txt
lightning
numpy
transformers
networkx
datasets
tqdm
scikit-learn
torchmetrics
ete3
lark
fire
langchain
llama-cpp-python
gradio==3.50